model matrix

+++ src/renderer/include/renderer/renderer.h
@@ -5,7 +5,7 @@ namespace wind {
 class Renderer {
 public:
     static void clear(ivec4 color);
-    static void drawRectangle(ivec4 rect, vec4 color);
+    static void drawRectangle(vec4 rect, vec4 color);
 };
 
 } // namespace wind
